+++ rskj-core/src/main/java/org/ethereum/crypto/altbn128/BN128G1.java
@@ -1,4 +1,4 @@
-package org.ethereum.crypto.altbn128.rust;
+package org.ethereum.crypto.altbn128;
 
 import java.math.BigInteger;
